pyproject.toml
[project]
name = "focus"
version = "0.1.0"
description = ""
packages = [{ include = "src" }]
authors = [

    {name = "Conor Grant",email = "conorlgrant@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"


dependencies = [
    "uvicorn (>=0.34.2,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "duckdb (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
